# Force early evaluation of index