--- a/hs/.gitignore
+++ b/hs/.gitignore
@@ -3,3 +3,5 @@
 dist-newstyle
 lcov.info
 bench/
+*.prof
+*.hp