In the previous implementation, every call to `wrappedLookup` rebuilt the `index` from scratch.

∆s from my machine are ~`10.98 s => 11.00 ms`

# Before

## Benchmarks

    time                 10.98 s    (9.956 s .. 12.26 s)
                         0.998 R²   (0.995 R² .. 1.000 R²)
    mean                 11.04 s    (10.78 s .. 11.27 s)
    std dev              301.4 ms   (113.2 ms .. 400.7 ms)

## Time Profile

    	   aoc +RTS -N -p -RTS run 3

    	total time  =        1.42 secs   (5691 ticks @ 1000 us, 4 processors)
    	total alloc = 16,177,748,224 bytes  (excludes profiling overheads)

    COST CENTRE         MODULE SRC                 %time %alloc

    wrappedLookup.index P03    src/P03.hs:17:5-59   90.7   94.6
    wrappedLookup.xmax  P03    src/P03.hs:18:5-47    8.9    5.3

# After

## Benchmarks

    time                 11.00 ms   (10.84 ms .. 11.14 ms)
                         0.998 R²   (0.996 R² .. 0.999 R²)
    mean                 11.06 ms   (10.97 ms .. 11.17 ms)
    std dev              263.3 μs   (178.7 μs .. 399.1 μs)

## Time Profile

    	   aoc +RTS -N -p -RTS run 3

    	total time  =        0.02 secs   (88 ticks @ 1000 us, 4 processors)
    	total alloc =  19,456,592 bytes  (excludes profiling overheads)

    COST CENTRE      MODULE                         SRC                                                 %time %alloc

    CAF              GHC.IO.Encoding                <entire-module>                                      38.6    0.0
    endOfLine        Data.Attoparsec.Text.Internal  Data/Attoparsec/Text/Internal.hs:444:1-69            31.8    0.0
    solution.\.atlas P03                            src/P03.hs:(8,7)-(10,54)                              8.0   46.6
    >>=.\            Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(148,9)-(149,44)    6.8   17.7
    MAIN             MAIN                           <built-in>                                            5.7    0.4
    >>=.\.succ'      Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:148:13-76           2.3   14.0
    readTextDevice   Data.Text.Internal.IO          libraries/text/Data/Text/Internal/IO.hs:133:39-64     1.1    0.0
    plus.\.lose'     Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:163:7-73            1.1    0.0
    >>=              Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(147,5)-(149,44)    1.1    0.0
    _1               Control.Lens.Tuple             src/Control/Lens/Tuple.hs:123:3-37                    1.1    5.0
    fmt              Santa.Runner                   src/Santa/Runner.hs:40:1-31                           1.1    0.0
    CAF              Santa.Main                     <entire-module>                                       1.1    0.0
    plus.\           Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(163,3)-(164,38)    0.0    2.2
    fmap.\.succ'     Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:173:11-58           0.0    2.9
    *>               Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:189:5-26            0.0    1.3
    path.loop        P03                            src/P03.hs:(22,5)-(24,14)                             0.0    1.1
    grid.tile        Import                         src/Import.hs:(37,5)-(41,22)                          0.0    3.0